added discount to unit tests

+++ Resourcerer_Sol/src/Resourcerer.DataAccess/Entities/CompositeSoldEvent.cs
@@ -3,7 +3,7 @@ public class CompositeSoldEvent : EntityBase
 {
     public double UnitsSold { get; set; }
     public double UnitPrice { get; set; }
-    public double TotalDiscountPercent { get; set; }
+    public int TotalDiscountPercent { get; set; }
 
     public Guid CompositeId { get; set; }
     public virtual Composite? Composite { get; set; }
