+++ Resourcerer_Sol/src/Resourcerer.Dtos/CompositeSoldEvents/CompositeSoldEventDto.cs
@@ -6,7 +6,7 @@ public class CompositeSoldEventDto : EntityDto
 {
     public double UnitsSold { get; set; }
     public double UnitPrice { get; set; }
-    public double TotalDiscountPercent { get; set; }
+    public int TotalDiscountPercent { get; set; }
 
     public Guid CompositeId { get; set; }
     public CompositeDto? Composite { get; set; }
