+++ Resourcerer_Sol/src/Resourcerer.Dtos/ElementPurchasedEvents/ElementPurchasedEventDto.cs
@@ -6,7 +6,7 @@ public class ElementPurchasedEventDto : EntityDto
 {
     public double UnitsBought { get; set; }
     public double UnitPrice { get; set; }
-    public double TotalDiscountPercent { get; set; }
+    public int TotalDiscountPercent { get; set; }
 
     public Guid ElementId { get; set; }
     public ElementDto? Element { get; set; }
