+++ Resourcerer_Sol/src/Resourcerer.Dtos/ElementSoldEvents/ElementSoldEventDto.cs
@@ -7,7 +7,7 @@ public class ElementSoldEventDto : EntityDto
 {
     public double UnitsSold { get; set; }
     public double PriceByUnit { get; set; }
-    public double TotalDiscountPercent { get; set; }
+    public int TotalDiscountPercent { get; set; }
 
     public UnitOfMeasureDto? UnitOfMeasure { get; set; }
